lua script

+++ store.go
@@ -6,5 +6,5 @@ import (
 
 type storage interface {
 	// Give a key and a expiration time, return value and ttl
-	Do(string, time.Duration) (int, float64, error)
+	Do(string, time.Duration) (int, int, error)
 }
